Connects a MapView and the Drink List View and database.
I have mashed all of the code together to create a working core program.
From this program we can start building things and making things look
nicer. There is plenty of stuff to do. I Bypassed the bar list view by
putting the map view in its place. The bar icons on the map will launch
an alert dialog and access the database and draw a listview of the
drinks according to what we query.