Made Changes to main map view. Added button.
This change makes the map use custom overlay items to allow us to
control how the markers are displayed. Custom overlay items will allow
us to display customized icons on the map. This commit also fixes some
of the layout of the bars so they are more visable and not overlapping
eachother so much. I also forced screen orientation to be vertical so
the database would not get recreated every time that the screen changes.
This commit also adds a button that will launch the barListView. The
button for the bar list view is temporary.