Implemented the call a cab button from Alan
I am commiting the call a cab button. I have many upgrades that i would
like to see done. Alan we will talk about this one and see what can be
improved. At the moment it does not work as you intended. This button is
going to work a little different than it is working now. We will base it
off of a random number. We will have more cab companies in there. I have
implemented the button the way that it is supposed to run all we have to
do is make it work correctly.