i am trying to pass data to the list view
I am working on passing data to the list view. I am building this from
the bottem up. So i am making sure that the list view can distribute the
data into the different days. because each list view is a fragment there
is no guarentee that all list views will be created if you try to pass
them data. To solve this i will provide a cache of data in the form of
an ArrayList <ArrayList <HashMap <String, String> > >   == Days <Drinks
<DrinkNames, DrinkDiscription. The DataStructure must be serializable to
be passed around in fragments so i must make a serializable class that
will contain the datastructure and be accessible from any fragment.