Fixed the libraries
Sorry i added the wrong one. The correct library that you should be
using is the sherlock maps r11. This library provided map fragment
support. Google just released maps support for fragments but i have not
had a chance to implement this yet. I am going to leave things the way
that they are until our project is over.